QuaternionArray.from_DCM: Handle undefined variable i.
The variable `i` was only defined when the program
entered one of the loops (for 'shepperd', 'itzhack',
or 'sarabandi' methods). If a RuntimeWarning
occurred in the 'hughes' or 'chiaverini' methods,
the variable `i` would have been undefined,
leading to an error. Now the code uses `idx`
which is defined in the outer scope as `None` by
default, and is ignored if not used in the error
message.

The variable `R` was also renamed to `rotmat`
to clarify that it represents a rotation matrix.